create proposals filter
