+++ apps/freelance/src/pages/Freelancer/SearchWork/index.tsx
@@ -1,6 +1,4 @@
 import { useState } from "react";
-import { useTranslation } from "react-i18next";
-import { SelectOptions } from "utils/select-options/options";
 import { Dashboard, Pagination, SearchWorkFilter, FilterBox } from "@freelance/components";
 import { JOBS_PAGE } from "src/utils/constants/breakpoint";
 import { useSearchWorkFormHook } from "./searchWorkFormHook";
@@ -8,22 +6,9 @@ import {
 	StyledPage,
 	InputWrapper,
 } from "./style";
-
-type user = "freelancer" | "employer";
-
-export interface IFormInput {
-	category: SelectOptions;
-	position: string;
-	skills: SelectOptions[];
-	employmentType: SelectOptions;
-	englishLevel: SelectOptions;
-	hourRate: SelectOptions;
-	availableAmountOfHour: SelectOptions;
-	typePage?: "proposals" | "job";
-}
+import { ROLES } from "src/utils/constants/roles";
 
 export function SearchWork() {
-	const user: user = "freelancer";
 	const {
 		onSubmit,
 		setFilter,
@@ -43,7 +28,7 @@ export function SearchWork() {
 		<StyledPage>
 			<Dashboard userRole="freelancer" typePage={JOBS_PAGE}>
 				<InputWrapper>
-					<Pagination itemsPerPage={5} user={user} data={filterJobs} typePage={JOBS_PAGE} />
+					<Pagination itemsPerPage={5} user={ROLES.FREELANCER} data={filterJobs} typePage={JOBS_PAGE} />
 				</InputWrapper>
 				<FilterBox isActive={isFilterOpen}>
 					<SearchWorkFilter
