+++ libs/components/src/lib/filter-container/filter-container.styled.tsx
@@ -5,16 +5,16 @@ interface IFilterContainer {
 }
 
 export const FilterContainer = styled.div<IFilterContainer>`
-    padding: 2rem;
+    padding: 1.5rem;
     border-radius: 15px;
     border:  1px solid ${({ theme }) => theme.colors.lightRed};
     position: absolute;
-    top: 16%;
+    top: 17%;
     right: 6%;
-    width: 30%;
+    width: 27%;
     background-color: ${({ theme }) => theme.colors.lightGrey};
     transition: 0.5s;
     transform: translateX(120%);
     
-    ${({ isShown }) => isShown && css`transform: translateX(20%)`}
+    ${({ isShown }) => isShown && css`transform: translateX(13%)`}
 `;
